Static analysis fixes and dependency version bumps

+++ izettle-cassandra/src/main/java/com/izettle/cassandra/DeterministicTimeUUIDFactory.java
@@ -68,7 +68,7 @@ public class DeterministicTimeUUIDFactory {
         // need to squeeze in type (4 MSBs in byte 6, clock hi)
         midhi &= ~0xF000; // remove high nibble of 6th byte
         midhi |= 0x1000; // type 1
-        long midhiL = (long) midhi;
+        long midhiL = midhi;
         midhiL = (midhiL << 32) >>> 32; // to get rid of sign extension
         // and reconstruct
         // last detail: must force 2 MSB to be '10'
